Fix tuple order for lon_err/lat_err in decode function
I noticed this when trying to get neighbors for the geohash 9g3m.
The current implementation was duplicating some neighbors and
returning an incorrect list.

It seems we accidentally had the x/y order flipped for the error
values returned from the decode function. I suspect this escaped
notice since the test cases we were working with had either very
small errors or were near enough to the equator that the x/y
errors were not significantly different.